Restructure frontend
