+++ frontend/src/app/components/mod.rs
@@ -1,0 +1,3 @@
+pub(in crate::app) use title::Title;
+
+mod title;
