+++ frontend/src/app/pages/landing.rs
@@ -1,0 +1,12 @@
+use yew::{Html, function_component, html};
+
+use crate::app::components::Title;
+
+#[function_component]
+pub(in crate::app) fn LandingPage() -> Html {
+    html! {
+        <>
+            <Title>{ "Welcome" }</Title>
+        </>
+    }
+}
