+++ frontend/src/app/pages/mod.rs
@@ -1,0 +1,12 @@
+pub(in crate::app) use admin::AdminPage;
+pub(in crate::app) use error::ErrorPage;
+pub(in crate::app) use landing::LandingPage;
+pub(in crate::app) use login::LoginPage;
+pub(self) use login::LoginQuery;
+pub(in crate::app) use logout::LogoutPage;
+
+mod admin;
+mod error;
+mod landing;
+mod login;
+mod logout;
