+++ frontend/src/app/utils.rs
@@ -1,6 +1,4 @@
 use gloo_net::http::Request;
-use yew::{Html, classes, function_component, html};
-use yew_autoprops::autoprops;
 
 use crate::net::responses;
 
@@ -69,13 +67,5 @@ pub(super) async fn get_current_user() -> Result<Option<state::User>, GetCurrent
             "Unexpected status code: {}",
             code
         ))),
-    }
-}
-
-#[autoprops]
-#[function_component]
-pub(super) fn Title(#[prop_or_default] children: &Html) -> Html {
-    html! {
-        <h1 class={ classes!("text-7xl", "text-center") }>{ children.clone() }</h1>
     }
 }
